+++ app/src/main/java/com/swadratna/swadratna_staff/data/remote/model/Category.kt
@@ -30,8 +30,8 @@ data class MenuItem(
     @SerializedName("category_id") val categoryId: Int,
     val name: String,
     val description: String,
-    val price: Int,
-    @SerializedName("discount_percentage") val discountPercentage: Int,
+    val price: Double,
+    @SerializedName("discount_percentage") val discountPercentage: Double,
     @SerializedName("discounted_price") val discountedPrice: Double, // Use Double since it's 151.2
     val currency: String,
     val image: String,
