+++ app/src/main/java/com/swadratna/swadratna_staff/data/remote/model/OrderX.kt
@@ -2,16 +2,16 @@ package com.swadratna.swadratna_staff.data.remote.model
 
 data class OrderX(
     val created_at: String,
-    val discount: Int,
+    val discount: Double,
     val id: Int,
     val location_id: Int,
     val order_date: String,
     val order_status: String,
-    val order_value: Int,
+    val order_value: Double,
     val restaurant_id: Int,
     val table_id: Int,
     val tenant_id: Int,
-    val total_value: Int,
+    val total_value: Double,
     val updated_at: String,
     val user_id: Int
 )
